# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SDM Invasion-Risk API",
    "description": "Species-keyed MLOps service for MaxEnt species distribution models: incremental GBIF/iNaturalist ingestion, conditional retraining, model lineage, map-layer descriptors (titiler / pg_tileserv) and automated bulletins.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/species/{taxon_key}/layers": {
      "get": {
        "tags": [
          "layers"
        ],
        "summary": "Get Layers",
        "operationId": "getLayers",
        "parameters": [
          {
            "name": "taxon_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Taxon Key"
            }
          },
          {
            "name": "model_version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LayerSet"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/species/{taxon_key}/bulletin": {
      "post": {
        "tags": [
          "bulletin"
        ],
        "summary": "Generate",
        "operationId": "generateBulletin",
        "parameters": [
          {
            "name": "taxon_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Taxon Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/BulletinRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "bulletin"
        ],
        "summary": "Bulletin Status",
        "operationId": "getBulletinStatus",
        "parameters": [
          {
            "name": "taxon_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Taxon Key"
            }
          },
          {
            "name": "model_version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulletinStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/species/{taxon_key}/bulletin/file": {
      "get": {
        "tags": [
          "bulletin"
        ],
        "summary": "Download",
        "operationId": "downloadBulletin",
        "parameters": [
          {
            "name": "taxon_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Taxon Key"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "pdf",
                "html"
              ],
              "type": "string",
              "default": "pdf",
              "title": "Format"
            }
          },
          {
            "name": "model_version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/pdf": {},
              "text/html": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job",
        "operationId": "getJob",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "List Jobs",
        "operationId": "listJobs",
        "parameters": [
          {
            "name": "taxon_key",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Taxon Key"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobOut"
                  },
                  "title": "Response Listjobs"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BulletinRequest": {
        "properties": {
          "model_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version"
          }
        },
        "type": "object",
        "title": "BulletinRequest"
      },
      "BulletinStatus": {
        "properties": {
          "taxon_key": {
            "type": "integer",
            "title": "Taxon Key"
          },
          "model_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version"
          },
          "pdf_available": {
            "type": "boolean",
            "title": "Pdf Available"
          },
          "html_available": {
            "type": "boolean",
            "title": "Html Available"
          },
          "pdf_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pdf Url"
          },
          "html_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Html Url"
          }
        },
        "type": "object",
        "required": [
          "taxon_key",
          "model_version",
          "pdf_available",
          "html_available"
        ],
        "title": "BulletinStatus"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobKind": {
        "type": "string",
        "enum": [
          "pipeline",
          "bulletin",
          "scenarios",
          "validation"
        ],
        "title": "JobKind"
      },
      "JobOut": {
        "properties": {
          "job_id": {
            "type": "string",
            "format": "uuid",
            "title": "Job Id"
          },
          "taxon_key": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Taxon Key"
          },
          "kind": {
            "$ref": "#/components/schemas/JobKind"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage"
          },
          "progress": {
            "type": "number",
            "title": "Progress"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "created_ts": {
            "type": "string",
            "format": "date-time",
            "title": "Created Ts"
          },
          "updated_ts": {
            "type": "string",
            "format": "date-time",
            "title": "Updated Ts"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "taxon_key",
          "kind",
          "status",
          "progress",
          "created_ts",
          "updated_ts"
        ],
        "title": "JobOut"
      },
      "JobStatus": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "succeeded",
          "failed"
        ],
        "title": "JobStatus"
      },
      "LayerCaveats": {
        "properties": {
          "model_type": {
            "type": "string",
            "enum": [
              "A",
              "B"
            ],
            "title": "Model Type"
          },
          "confidence_label": {
            "type": "string",
            "enum": [
              "standard",
              "lower"
            ],
            "title": "Confidence Label"
          },
          "transferability_caveat": {
            "type": "string",
            "title": "Transferability Caveat"
          },
          "mess_required": {
            "type": "boolean",
            "title": "Mess Required",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "model_type",
          "confidence_label",
          "transferability_caveat"
        ],
        "title": "LayerCaveats"
      },
      "LayerSet": {
        "properties": {
          "taxon_key": {
            "type": "integer",
            "title": "Taxon Key"
          },
          "model_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version"
          },
          "occurrences": {
            "$ref": "#/components/schemas/VectorLayer"
          },
          "native_range": {
            "$ref": "#/components/schemas/VectorLayer"
          },
          "suitability": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RasterLayer"
              },
              {
                "type": "null"
              }
            ]
          },
          "mess": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RasterLayer"
              },
              {
                "type": "null"
              }
            ]
          },
          "zones": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RasterLayer"
              },
              {
                "type": "null"
              }
            ]
          },
          "extrapolation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RasterLayer"
              },
              {
                "type": "null"
              }
            ],
            "description": "MESS < 0 mask; always offered as an overlay with any projection"
          },
          "scenarios": {
            "items": {
              "$ref": "#/components/schemas/ScenarioLayer"
            },
            "type": "array",
            "title": "Scenarios"
          },
          "scenarios_available": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Scenarios Available",
            "description": "Scenarios registered for the model's bioclim version (projected or not)"
          },
          "caveats": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LayerCaveats"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "taxon_key",
          "model_version",
          "occurrences",
          "native_range"
        ],
        "title": "LayerSet"
      },
      "LegendEntry": {
        "properties": {
          "value": {
            "type": "number",
            "title": "Value"
          },
          "color": {
            "type": "string",
            "title": "Color"
          },
          "label": {
            "type": "string",
            "title": "Label"
          }
        },
        "type": "object",
        "required": [
          "value",
          "color",
          "label"
        ],
        "title": "LegendEntry"
      },
...
      "RasterLayer": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "tile_url": {
            "type": "string",
            "title": "Tile Url",
            "description": "XYZ template with {z}/{x}/{y}"
          },
          "kind": {
            "type": "string",
            "enum": [
              "continuous",
              "categorical"
            ],
            "title": "Kind"
          },
          "legend": {
            "items": {
              "$ref": "#/components/schemas/LegendEntry"
            },
            "type": "array",
            "title": "Legend"
          },
          "description": {
            "type": "string",
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "tile_url",
          "kind",
          "legend",
          "description"
        ],
        "title": "RasterLayer"
      },
...
      "ScenarioLayer": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "'scenario:<name>'"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "suitability": {
            "$ref": "#/components/schemas/RasterLayer"
          },
          "extrapolation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RasterLayer"
              },
              {
                "type": "null"
              }
            ],
            "description": "This scenario's own MESS < 0 mask (null for legacy projections)"
          },
          "extrapolated_land_fraction": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extrapolated Land Fraction"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "label",
          "suitability"
        ],
        "title": "ScenarioLayer"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "VectorLayer": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "tile_url": {
            "type": "string",
            "title": "Tile Url",
            "description": "MVT XYZ template; extra query params may be appended"
          },
          "source_layer": {
            "type": "string",
            "title": "Source Layer"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "tile_url",
          "source_layer"
        ],
        "title": "VectorLayer"
      }
    }
  }
}